pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Testing_Integration"
description = "TPU-Tester bench verify that TPU's are working bevor leaving production"
readme = "README.md"
requires-python = ">=3.10"
classifiers = ["Programming Language :: Python :: 3"]
dynamic = ['version']
dependencies = [
    "IPython",
    "paramiko",
    "pyserial",
    "tkinter",
]

[project.scripts]
tpu_tester = "main:main"

[tool.hatch.version]
path = "tpu_tester/version.py"

[tool.hatch.build]
only-include = ["tpu_tester"]

[tool.hatch.envs.default]
python = "3.10"
dependencies = [
    "ruff",
    "pytest",
]

[tool.hatch.envs.default.scripts]
lint = ["ruff check .", "ruff clean"]
test = ["pytest . --cache-clear"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]
